requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
langchain-classic
python-dotenv
gradio
pydantic
tavily-python
langsmith